+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 2.0.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Now you can customize tipologie_persona from the control panel.
+  [arsenico13]
 
 
 2.0.1 (2021-03-24)
@@ -15,6 +16,7 @@ Changelog
   [cekk]
 - Add remoteUrl to summarize serialization for Link content-type.
   [cekk]
+
 
 2.0.0 (2021-03-02)
 ------------------
@@ -47,6 +49,7 @@ Changelog
 - Do not run dependencies when upgrading plone.app.registry.
   [cekk]
 
+
 1.0.6 (2021-02-15)
 ------------------
 
@@ -61,6 +64,7 @@ Changelog
   [deodorhunter]
 - Remove reference limit in "persone_struttura" field.
   [cekk]
+
 
 1.0.4 (2021-02-05)
 ------------------
@@ -84,6 +88,7 @@ Changelog
   [cekk]
 - Do not break *eventoCreateHandler* when copying and event.
   [cekk]
+
 
 1.0.1 (2020-12-14)
 ------------------